+++ src/app/cart/models/cart.model.ts
@@ -1,9 +1,9 @@
-export class ProductItem{
+export class CartItem{
 
     constructor(
         public id: number = 0,
         public name: string,
         public price: number = 0,
-        public isPresent: boolean = false
+        public quantity: number = 0
     ){}
 }