+++ src/app/cart/services/cart-communicator.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CartCommunicatorService } from './cart-communicator.service';
+
+describe('CartCommunicatorService', () => {
+  let service: CartCommunicatorService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CartCommunicatorService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});