+++ src/app/products/components/product/product.component.ts
@@ -1,4 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, Output, OnInit } from '@angular/core';
+
+import { ProductItem } from '../../models/product.model';
 
 @Component({
   selector: 'app-product',
@@ -6,14 +8,26 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./product.component.css']
 })
 export class ProductComponent implements OnInit {
+  
+  @Input()
+  products: Array<ProductItem> =[];
 
-  constructor() { }
+  @Output()
+  productToBuy: EventEmitter<ProductItem> = new EventEmitter<ProductItem>();
+
+  headers: string[] = ['id', 'name', 'price', 'actions'];
 
   ngOnInit(): void {
+    
   }
 
   onBuy(): void {
     console.log('Item purchased!');
+  }
+
+  onAddInCart(product: ProductItem): void {
+   
+    this.productToBuy.emit(product);
   }
 
 }