+++ src/app/products/services/products.service.ts
@@ -11,9 +11,9 @@ export class ProductsService {
 
   getProducts(): Array<ProductItem>{
     return[
-      new ProductItem(1, 'Product1', 11),
-      new ProductItem(2, 'Product2', 22),
-      new ProductItem(3, 'Product3', 33),
+      new ProductItem(1, 'Product1', 11, true),
+      new ProductItem(2, 'Product2', 22, false),
+      new ProductItem(3, 'Product3', 33, true),
     ];
   }
 }